Fix some packets still writing in bytes and not being read properly. Will fix quite a few bugs.

+++ src/minecraft/micdoodle8/mods/galacticraft/core/network/GCCorePacketDimensionListPlanets.java
@@ -26,7 +26,7 @@ public class GCCorePacketDimensionListPlanets implements IGalacticraftAdvancedPa
 
         try
         {
-        	data.writeByte(GCCorePacketDimensionListPlanets.packetID);
+        	data.writeInt(GCCorePacketDimensionListPlanets.packetID);
         	data.writeInt(col.size());
             final Iterator var3 = col.iterator();
